requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
confusable-homoglyphs==3.2.0
distlib==0.3.3
Django==3.2.9
django-debug-toolbar==3.2.4
django-filter==21.1
django-registration==3.2
djangorestframework==3.13.1
filelock==3.4.0
Markdown==3.3.6
platformdirs==2.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
virtualenv==20.10.0
